+++ test/support/image_plug/imgproxy_wire_conformance_test/origin_image.ex
@@ -1,0 +1,13 @@
+defmodule ImgproxyWireConformanceTest.OriginImage do
+  @moduledoc false
+
+  use Boundary, top_level?: true, deps: []
+
+  def call(conn, _opts) do
+    body = File.read!("priv/static/images/beach.jpg")
+
+    conn
+    |> Plug.Conn.put_resp_content_type("image/jpeg")
+    |> Plug.Conn.send_resp(200, body)
+  end
+end
