+++ test/support/image_plug/imgproxy_wire_conformance_test/origin_should_not_fetch.ex
@@ -1,0 +1,7 @@
+defmodule ImgproxyWireConformanceTest.OriginShouldNotFetch do
+  @moduledoc false
+
+  use Boundary, top_level?: true, deps: []
+
+  def call(_conn, _opts), do: raise("origin should not fetch")
+end
